backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
dj-rest-auth==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
# Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0 